Introduction to server side fetching
Kind of opposite of server side fetching, it load all the data from the server, an then give back the page when everything is fetched. The disavantage is, if the data take too many time to load, it cause a long waiting on client side, giving an bad browser experiance.