Introduction to dynamic component imports (Lazy load of components)
Used on components that dont are always used by the user, so it only get imported when it is used. It also give the possibility to dont use ssr, witch means server side rendering, if the component uses global varibales like document or window of the browser, since all the components on next are generally rendered at server.