Local variables with Next Js
In Next JS, you can add multiples types of local varibles: .env.development: for development enviroment, .env.production and .env.test. All this .env files goes to github, it is genral files that dont uses any important data on it. If that is need to mantain a sensitive file with passwords and etc., you can use the .env.local file, that does not go to github by default.
Also, in Next, envirioment variables are only available on Server side, so, browser cant access it by default. To make it, you need to name it with the prefix NEXT_PUBLIC_, like:
NEXT_PUBLIC_API_URL.
That way, you can get the data even on client side.